+++ docs/source/whatsnew.rst
@@ -5,6 +5,7 @@ What's New
 
 v4.3.4
 ------
+- :meth:`Device.set_layout()` to allow a :class:`DeviceLayout` to be configured on a device
 - New method :meth:`Workspace.set_professional_license()` allows Workspaces with a Workspace license to be converted to a Professional license. Not that, as of this release, the Webex API does now allow a Professional Workspace to be downgraded to a Workspace license.
 - New property :attr:`Workspace.license_type` returns 'PROFESSIONAL' or 'WORKSPACE' to allow the license type to be determined
 - :meth:`WorkspaceList.create()` now accepts a ``license_type`` argument to allow Professional licenses to be assigned to a Workspace